Creating agent-log-pipeline e2e tests
