+++ .gitlab-ci.yml
@@ -887,8 +887,6 @@ workflow:
         - comp/aggregator/**/*
         - comp/collector/**/*
         - comp/core/agenttelemetry/**/*
-        - comp/core/autodiscovery/providers/config_reader*.go
-        - comp/core/autodiscovery/providers/file*.go
         - comp/dogstatsd/**/*
         - comp/forwarder/**/*
         - comp/serializer/**/*
@@ -902,33 +900,17 @@ workflow:
         - rtloader/**/*
       compare_to: main # TODO: use a variable, when this is supported https://gitlab.com/gitlab-org/gitlab/-/issues/369916
 
-.on_aml_or_e2e_changes:
+.on_alp_or_e2e_changes:
   - !reference [.on_e2e_main_release_or_rc]
   - changes:
       paths:
-        - test/new-e2e/tests/agent-metrics-logs/**/*
-        - cmd/agent/subcommands/dogstatsd*/*
+        - test/new-e2e/tests/agent-log-pipelines/**/*
         - cmd/agent/subcommands/streamlogs/*
-        - cmd/dogstatsd/**/*
-        - comp/agent/jmxlogger/**/*
-        - comp/aggregator/**/*
-        - comp/collector/**/*
         - comp/core/agenttelemetry/**/*
         - comp/core/autodiscovery/providers/config_reader*.go
         - comp/core/autodiscovery/providers/file*.go
-        - comp/dogstatsd/**/*
-        - comp/forwarder/**/*
         - comp/logs/**/*
-        - comp/serializer/**/*
-        - pkg/aggregator/**/*
-        - pkg/collector/**/*
-        - pkg/commonchecks/**/*
-        - pkg/jmxfetch/**/*
         - pkg/logs/**/*
-        - pkg/metrics/**/*
-        - pkg/persistentcache/**/*
-        - pkg/serializer/**/*
-        - rtloader/**/*
       compare_to: main # TODO: use a variable, when this is supported https://gitlab.com/gitlab-org/gitlab/-/issues/369916
 
 .on_cws_or_e2e_changes:
