+++ test/new-e2e/tests/agent-log-pipelines/k8s-logs/file_tailing_test.go
@@ -9,6 +9,7 @@ import (
 	"context"
 	_ "embed"
 	"fmt"
+	kindfilelogger "github.com/DataDog/datadog-agent/test/new-e2e/tests/agent-log-pipelines/kindfilelogging"
 	"testing"
 	"time"
 
@@ -20,7 +21,6 @@ import (
 
 	"github.com/DataDog/datadog-agent/test/new-e2e/pkg/e2e"
 	"github.com/DataDog/datadog-agent/test/new-e2e/pkg/environments"
-	kindfilelogger "github.com/DataDog/datadog-agent/test/new-e2e/tests/agent-metrics-logs/kindfilelogging"
 )
 
 type k8sSuite struct {
