+++ test/new-e2e/tests/agent-log-pipelines/kindfilelogging/kind.go
