+++ test/new-e2e/tests/agent-log-pipelines/linux-log/file-tailing/config/utf-16-le.yaml
