+++ test/new-e2e/tests/agent-log-pipelines/linux-log/file-tailing/multi_line_test.go
@@ -8,6 +8,7 @@ package linuxfiletailing
 import (
 	_ "embed"
 	"fmt"
+	"github.com/DataDog/datadog-agent/test/new-e2e/tests/agent-log-pipelines/utils"
 	"testing"
 	"time"
 
@@ -18,7 +19,6 @@ import (
 	"github.com/DataDog/datadog-agent/test/new-e2e/pkg/e2e"
 	"github.com/DataDog/datadog-agent/test/new-e2e/pkg/environments"
 	awshost "github.com/DataDog/datadog-agent/test/new-e2e/pkg/provisioners/aws/host"
-	"github.com/DataDog/datadog-agent/test/new-e2e/tests/agent-metrics-logs/log-agent/utils"
 )
 
 //go:embed config/automulti.yaml
