+++ test/new-e2e/tests/agent-log-pipelines/linux-log/journald/log-config/include.yaml
