+++ test/new-e2e/tests/agent-log-pipelines/listener/testdata/agent-logs-udp-conf.yaml
