+++ test/new-e2e/tests/agent-log-pipelines/listener/testdata/tcp-compose.yaml
