+++ test/new-e2e/tests/agent-log-pipelines/listener/testdata/udp-compose.yaml
