+++ test/new-e2e/tests/agent-log-pipelines/windows-log/file-tailing/log-config/config.yaml
